:truck: move `is_applicable_for_billing` method, and move tests to test_billing_manager module
